WIP continuing the refactor
