+++ engine/launcher/src/test/scala/org/enso/launcher/NativeTest.scala
@@ -1,6 +1,6 @@
 package org.enso.launcher
 
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 
 import java.nio.file.{Files, Path}
 import org.enso.runtimeversionmanager.test.NativeTestHelper
