+++ lib/scala/distribution-manager/src/main/scala/org/enso/distribution/OS.scala
@@ -1,4 +1,4 @@
-package org.enso.distributionmanager
+package org.enso.distribution
 
 import com.typesafe.scalalogging.Logger
 import io.circe.{Decoder, DecodingFailure}
