+++ lib/scala/distribution-manager/src/main/scala/org/enso/distribution/locking/ResourceManager.scala
@@ -1,4 +1,4 @@
-package org.enso.runtimeversionmanager.locking
+package org.enso.distribution.locking
 
 import com.typesafe.scalalogging.Logger
