+++ lib/scala/runtime-version-manager-test/src/main/scala/org/enso/runtimeversionmanager/test/NativeTestHelper.scala
@@ -1,6 +1,6 @@
 package org.enso.runtimeversionmanager.test
 
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 
 import java.io.{
   BufferedReader,
