+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/components/Manifest.scala
@@ -5,7 +5,7 @@ import java.nio.file.Path
 import cats.Show
 import io.circe.{yaml, Decoder}
 import nl.gn0s1s.bump.SemVer
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 import org.enso.pkg.SemVerJson._
 import org.enso.runtimeversionmanager.components.Manifest.{
   JVMOption,
