+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/distribution/DistributionManager.scala
@@ -2,7 +2,7 @@ package org.enso.runtimeversionmanager.distribution
 
 import java.nio.file.{Files, Path}
 import com.typesafe.scalalogging.Logger
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 import org.enso.runtimeversionmanager.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.{Environment, FileSystem}
