+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/locking/Resources.scala
@@ -1,24 +1,10 @@
 package org.enso.runtimeversionmanager.locking
 
 import nl.gn0s1s.bump.SemVer
+import org.enso.distribution.locking.Resource
 import org.enso.runtimeversionmanager.components.GraalVMVersion
 
-/** Represents a resource that can be locked.
-  */
-trait Resource {
-
-  /** Name of the resource.
-    *
-    * Must be a valid filename part.
-    */
-  def name: String
-
-  /** A message that is displayed by default if the lock on that resource cannot
-    * be acquired immediately.
-    */
-  def waitMessage: String
-}
-object Resource {
+object Resources {
 
   /** Synchronizes launcher upgrades.
     */
