+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/releases/EnsoReleaseProvider.scala
@@ -1,6 +1,6 @@
 package org.enso.runtimeversionmanager.releases
 import nl.gn0s1s.bump.SemVer
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 
 import scala.util.{Failure, Success, Try}
