+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/releases/testing/TestArchivePackager.scala
@@ -1,6 +1,6 @@
 package org.enso.runtimeversionmanager.releases.testing
 
-import org.enso.distributionmanager.OS
+import org.enso.distribution.OS
 
 import java.nio.file.Path
 import org.enso.runtimeversionmanager.FileSystem
